requirements.txt
joblib==1.0.0
matplotlib==3.3.1
mpmath==1.1.0
networkx==2.5
numpy==1.19.1
observations==0.1.4
pandas
pillow==8.0.1
pydot==1.4.1
scikit-learn==0.23.2
scipy==1.5.2
setproctitle==1.1.10
scikit-image==0.17.2
spectral==0.22.1
statsmodels==0.12.1
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
tqdm==4.54.1
